sync
Former-commit-id: 3d8a7a4206a0b1bc8c6bcec89f4fcbf406717326 [formerly 41c92eb92f0a0c27170cae5ad009dd58919a1fae]
Former-commit-id: 9536edfcb09b316b3e5c16a039699fae39eeb16d
Former-commit-id: 9a3e8634555dc93b2ab481b5b328b2c206848f1b [formerly 85d4b5979137a9d12a80e1155fb8428a5fca9558] [formerly ca5ccb136a998e786b3c69db2c898d7841868358 [formerly 3134e60adf88c1efdb20b49af4737413a7878b96]] [formerly ba279b1c3611e788702712c07547df313ac7aa6a [formerly b7b65dcddf8e12abdf0a525b0efec478c0c19173] [formerly 7653ec66cd68e4a5c3d1d1452ce413d78ab664db [formerly eb5c28dea94815bfacfe41a7a4b82981e3882b66]]]
Former-commit-id: 58a8de234d793666ad547fcec573fe2dcd88a53c [formerly c766f76e1fd8434ad722d31fee8625b3282aff7a] [formerly 179dc59075d2f55ffb41f075d50392000e5bd869 [formerly f284a7c836e414b10473d64d4446017125d066b5]]
Former-commit-id: 30079600f204f92746e740fd48e387a51e6b061d [formerly 7f580d52e44e102bd730a0ffcf4c2e94547be165]
Former-commit-id: 3277d1f14e9696871f7562922579aac603567c34

+++ src/demand_generator.hpp
@@ -15,7 +15,7 @@ class DemandGenerator {
   public:
     /// \brief Constructor.
     explicit DemandGenerator(std::string _path_to_taxi_data, std::string _simulation_start_time,
-                             double _request_density);
+                             float _request_density);
 
     /// \brief Main functor that generates the requests til the target system time.
     std::vector<Request> operator()(uint64_t target_system_time_ms);
